backend/requirements.txt
Flask
flask-cors
SQLAlchemy
psycopg2-binary  # For PostgreSQL
requests
python-dotenv
